backend/requirements.txt
# BusPassPro - Backend Dependencies

# Django Core
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2

# Async / WebSockets
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Database
psycopg2-binary==2.9.10
django-redis==5.4.0

# Auth & Security
PyJWT==2.9.0

# Payments
razorpay==1.4.2

# QR Code
qrcode[pil]==8.0
pyzbar==0.1.9
Pillow==11.1.0

# AI / ML
scikit-learn==1.6.1
numpy==2.2.3
networkx==3.4.2
joblib==1.4.2
python-dateutil==2.9.0

# Utilities
python-decouple==3.8
celery==5.4.0
redis==5.2.1
